Controller::Release::find - move data handling to Document module
Keep the controller code concise and unaware of the data structure.

+++ lib/MetaCPAN/Server/Controller/Release.pm
@@ -22,13 +22,8 @@ __PACKAGE__->config(
 sub find : Path('') : Args(1) {
     my ( $self, $c, $name ) = @_;
     my $file = $self->model($c)->raw->find($name);
-    if ( !defined $file ) {
-        $c->detach( '/not_found', [] );
-    }
-    $c->stash( $file->{_source}
-            || single_valued_arrayref_to_scalar( $file->{fields} ) )
-        || $c->detach( '/not_found',
-        ['The requested field(s) could not be found'] );
+    $c->detach( '/not_found', [] ) unless $file;
+    $c->stash($file);
 }
 
 sub get : Path('') : Args(2) {
